requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyyaml>=6.0.1
streamlit>=1.28.0
moviepy>=1.0.3
python-dotenv>=1.0.0 